refactor: Extract DhtServer in separate file
