+++ CMakeLists.txt
@@ -230,6 +230,8 @@ set(${PROJECT_NAME}_SOURCES
   src/core/core.h
   src/core/corestructs.cpp
   src/core/corestructs.h
+  src/core/dhtserver.cpp
+  src/core/dhtserver.h
   src/core/indexedlist.h
   src/core/recursivesignalblocker.cpp
   src/core/recursivesignalblocker.h
