+++ src/core/dhtserver.cpp
@@ -1,0 +1,41 @@
+/*
+    Copyright © 2017 by The qTox Project Contributors
+
+    This file is part of qTox, a Qt-based graphical interface for Tox.
+
+    qTox is libre software: you can redistribute it and/or modify
+    it under the terms of the GNU General Public License as published by
+    the Free Software Foundation, either version 3 of the License, or
+    (at your option) any later version.
+
+    qTox is distributed in the hope that it will be useful,
+    but WITHOUT ANY WARRANTY; without even the implied warranty of
+    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+    GNU General Public License for more details.
+
+    You should have received a copy of the GNU General Public License
+    along with qTox.  If not, see <http://www.gnu.org/licenses/>.
+*/
+
+#include "dhtserver.h"
+
+/**
+ * @brief   Compare equal operator
+ * @param   other   the compared instance
+ * @return  true, if equal; false otherwise
+ */
+bool DhtServer::operator==(const DhtServer& other) const
+{
+    return this == &other || (port == other.port && address == other.address
+                              && userId == other.userId && name == other.name);
+}
+
+/**
+ * @brief   Compare not equal operator
+ * @param   other   the compared instance
+ * @return  true, if not equal; false otherwise
+ */
+bool DhtServer::operator!=(const DhtServer& other) const
+{
+    return !(*this == other);
+}
